+++ static/app/gettingStartedDocs/android/android/onboarding.spec.tsx
@@ -2,7 +2,7 @@ import {renderWithOnboardingLayout} from 'sentry-test/onboarding/renderWithOnboa
 import {screen} from 'sentry-test/reactTestingLibrary';
 import {textWithMarkupMatcher} from 'sentry-test/utils';
 
-import docs from './android';
+import docs from '.';
 
 describe('android onboarding doc', () => {
   it('renders doc correctly', async () => {
